apps/backend/requirements.txt
fastapi
uvicorn
psutil
opentelemetry-distro
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc
prometheus-client
pydantic